pyproject.toml
[project]
name = "azure-translator-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-translation-text>=1.0.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
